benchmark-mlflow/requirements.txt
mlflow>=2.0.0
requests>=2.28.0
PyYAML>=6.0
python-dotenv>=1.0.0
pydantic>=1.10.0
fastapi>=0.104.1
uvicorn>=0.24.0 